CodeCoa Room 1
I spent the majority of my time just figuring out how the controls worked, but I found my way around and quickly made the start of my room. Its very basic but for a rough start with no theme yet i think i did pretty good. I used the bevel extrude and combine tool a lot. I had a lot of trouble with the legs and the ears of the chair cushion. I couldn't figure out how to make it quite like i wanted, but in the jankiness of it all i think that added some personality to it. For example the bottom of the chair when smoothed looks almost too smooth so i kept its sharp edges. Maybe later I can make it look soft but keep its weird shape.