Books, Boxes, Frames, PictureFrame, More Lighting
Using the bevel, extrude, and booleans I made some organic looking cardboard boxes, and fixed up the shelves on the side of the room. Then I made the lighting better and added a picture frame on the wall.
next time i will make the lantern and possibly the couch. I also need to add stuff to the bookshelves