The Bog Part 2
second part. updated the building and most textures to be more coherent and get their point across, worked with uv's more, and then began the start of the main part of the bog which will branch off and spill outwards like an overgrowth of sorts. everything looks great when rendered, but i have no lighting so that is what i will be working on next time.